+++ src/repository/IngredientRepository.java
@@ -26,7 +26,7 @@ public class IngredientRepository implements Repository<Ingredient> {
             id,
             rs.getString("name"),
             rs.getTimestamp("update_datetime").toLocalDateTime(),
-            rs.getBigDecimal("unit_price"),
+            rs.getDouble("unit_price"),
             Unit.valueOf(rs.getString("unit")),
             priceHistories
         );
@@ -116,11 +116,11 @@ public class IngredientRepository implements Repository<Ingredient> {
          """;
         try{
             PreparedStatement prs = connection.prepareStatement(query);
-            prs.setString (1, toCreate.getId());
-            prs.setString (2, toCreate.getName());
+            prs.setString(1, toCreate.getId());
+            prs.setString(2, toCreate.getName());
             prs.setTimestamp(3, Timestamp.valueOf(toCreate.getUpdateDatetime()));
-            prs.setBigDecimal (4, toCreate.getUnitPrice());
-            prs.setString (5, toCreate.getUnit().toString());
+            prs.setDouble(4, toCreate.getUnitPrice());
+            prs.setString(5, toCreate.getUnit().toString());
             prs.executeUpdate();
             return this.findById(toCreate.getId());
         }catch (SQLException error){
@@ -140,11 +140,11 @@ public class IngredientRepository implements Repository<Ingredient> {
         """;
         try{
             PreparedStatement prs = connection.prepareStatement(query);
-            prs.setString (1, toUpdate.getName());
+            prs.setString(1, toUpdate.getName());
             prs.setTimestamp(2, Timestamp.valueOf(toUpdate.getUpdateDatetime()));
-            prs.setBigDecimal(3, toUpdate.getUnitPrice());
-            prs.setString (4, toUpdate.getUnit().toString());
-            prs.setString (5, toUpdate.getId());
+            prs.setDouble(3, toUpdate.getUnitPrice());
+            prs.setString(4, toUpdate.getUnit().toString());
+            prs.setString(5, toUpdate.getId());
             prs.executeUpdate();
             return this.findById(toUpdate.getId());
         }catch (SQLException error){
